bug fix for sniper zoom breaking death menu.
Sniper hud overlay stayed up when dying while zoomed in. This meant player could not click any of the buttons on the menu, as the overlay was present over top of it. I created a ForceZoomOut() function in playerController, and call it during death function in gameManager. Now if player dies while zoomed, first the HUD is restored, camera unzoomed, and sniper overlay is hidden, before the rest of the death function plays out.